repository test 추가;

+++ src/main/java/com/example/file_extension_blocker/repository/DefaultExtensionRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.file_extension_blocker.entity.DefaultFileExtension;
 
-public interface DefaultExtensionRepository extends JpaRepository<DefaultFileExtension, Integer> {
+public interface DefaultExtensionRepository extends JpaRepository<DefaultFileExtension, Long> {
 
 	Boolean existsByName(String name);
 }
